更新查询直属会员，游戏余额检查bug修复

+++ mxsohaGameServer/src/service/UserService.java
@@ -116,7 +116,7 @@ public interface UserService {
 	 * @param userId
 	 * @return
 	 */
-	int getUserMoneyById(int userId);
+	double getUserMoneyById(int userId);
 	
 
 }
